Refactor Card, Improve efficiency of list and edit
The list card has been split up into pieces to more
easily rearrange and develop the features at each
point in the card.

The list is now much more efficient and does not
require a page refresh on mutation. I'm using a
setTimeout which may not be a good idea, but it
seems to be working in development pretty well.