Modify Card, Details, and FormViewModel behavior
To avoid breaking the roolz of hooks I changed the
instantiation of the FormViewModel to be encapsulated
into the function `useFormViewModel`, which also allows
me to move the trpc calls into that function instead
of in the dialogs, which is a nice benefit.

	I think using the queries directly in the components
	has an effect that is separate and better than using
	useState with the data, more experiments are needed
	with respect to that though.

Improved the behavior of the detail page so that
everything refreshes correctly and started to implement
the color more throughout the details page as well as the
icon.

Lots of style based work on the cards. Moved tags and
notes into an accordion with plans to expand into more
items through there.